Исправил ошибку неправильной длины при парсинге пакета со счетчиками Изменил css стиль кнопки sign out. Добавил приведение к типу double при приращении фонда приза

+++ Server/Prize/ServerPrize.cs
@@ -49,7 +49,7 @@ namespace Server.Prize
                     if (!groupPrize.WaitConfirm)
                     {
                         groupPrize.LastTerminalId = terminalId;
-                        var value = input * groupPrize.Percent / 100;
+                        var value = (double)(input * groupPrize.Percent / 100);
                         groupPrize.AddInFund(value);
                     }
                 }
